Updates msim builder to provide column numbers
Adds functionality to the modelsim builder to provide column numbers for
errors and warnings. Analyzes the error message from modelsim, searches
for a token in quotes, and finds that in the source file line. Adds the
necessary function signature to the ghdl builder and the change for
column_start instead of column_number